+++ enpt/model/images/images_sensorgeo.py
@@ -352,6 +352,77 @@ class EnMAP_Detector_SensorGeo(_EnMAP_Image):
                 self.logger.warning("Could not save the %s attribute '%s' as it does not exist in the current image."
                                     % (self.detector_name, attrName))
 
+    def _transform_raster_geometry_from_other_detector(self,
+                                                       array: np.ndarray,
+                                                       src_lons: np.ndarray,
+                                                       src_lats: np.ndarray,
+                                                       src_epsg: int,
+                                                       resamp_alg: str = 'nearest',
+                                                       respect_keystone: bool = False
+                                                       ) -> np.ndarray:
+        """Transform the given input raster from SWIR to VNIR or from SWIR to VNIR sensor geometry.
+
+        NOTE: - The transformation target is always the EnMAP_Detector_SensorGeo instance sensor geometry
+                (e.g., VNIR sensorgeo if self.detector_name == 'VNIR').
+              - In case a 3D array is given and the array has the exact dimensions of the source detector,
+                the full geolayer is only used if 'respect_keystone' is set to True. This saves computation time
+                for input arrays where the keystone uncertainty does not matter.
+
+
+        :param array:               input array to be transformed (2- or 3-dimensional)
+        :param src_lons:            geolayer longitudes corresponding to the input array
+                                    (same nbands like the source detector)
+        :param src_lats:            geolayer latitudes corresponding to the input array
+                                    (same nbands like the source detector)
+        :param src_epsg:            projection EPSG code of the source array
+        :param resamp_alg:          resampling algorithm ('nearest', 'bilinear', 'gauss', 'custom')
+        :param respect_keystone:    whether to use the full geoarray (all bands) in case a 3D array
+                                    in the dimension of the source detector is passed (default: False)
+        :return:
+        """
+        detN = self.detector_name
+        if self.detector_meta.lons is None or self.detector_meta.lats is None:
+            raise RuntimeError(f"The {detN} geolayer must be computed first "
+                               f"to transform arrays from {'SWIR' if detN == 'VNIR' else 'VNIR'} "
+                               f"to {detN} sensor geometry.")
+
+        if src_lons.shape != src_lats.shape:
+            raise ValueError("'src_lons' must have the same shape as 'src_lats'.")
+
+        vnir_lons = self.detector_meta.lons if detN == 'VNIR' else src_lons
+        vnir_lats = self.detector_meta.lats if detN == 'VNIR' else src_lats
+        swir_lons = self.detector_meta.lons if detN == 'SWIR' else src_lons
+        swir_lats = self.detector_meta.lats if detN == 'SWIR' else src_lats
+        prj_vnir = self.detector_meta.epsg_ortho if detN == 'VNIR' else src_epsg
+        prj_swir = self.detector_meta.epsg_ortho if detN == 'SWIR' else src_epsg
+
+        # use first geolayer band if the input array has only one band
+        if array.ndim == 2 or \
+           array.shape[2] != src_lons.shape[2] or \
+           not respect_keystone:
+            vnir_lons = vnir_lons[:, :, 0]
+            vnir_lats = vnir_lats[:, :, 0]
+            swir_lons = swir_lons[:, :, 0]
+            swir_lats = swir_lats[:, :, 0]
+
+        from ...processors.spatial_transform import VNIR_SWIR_SensorGeometryTransformer
+        VS_SGT = VNIR_SWIR_SensorGeometryTransformer(lons_vnir=vnir_lons,
+                                                     lats_vnir=vnir_lats,
+                                                     lons_swir=swir_lons,
+                                                     lats_swir=swir_lats,
+                                                     prj_vnir=prj_vnir,
+                                                     prj_swir=prj_swir,
+                                                     res_vnir=(30, 30),
+                                                     res_swir=(30, 30),
+                                                     resamp_alg=resamp_alg,
+                                                     # radius_of_influence=45,
+                                                     nprocs=self.cfg.CPUs
+                                                     )
+        if detN == 'VNIR':
+            return VS_SGT.transform_sensorgeo_SWIR_to_VNIR(array)
+        else:
+            return VS_SGT.transform_sensorgeo_VNIR_to_SWIR(array)
+
 
 class EnMAP_VNIR_SensorGeo(EnMAP_Detector_SensorGeo):
     def __init__(self, root_dir: str, config: EnPTConfig, logger=None, meta=None) -> None:
@@ -387,6 +458,27 @@ class EnMAP_VNIR_SensorGeo(EnMAP_Detector_SensorGeo):
         if self.paths.mask_cirrus:
             self.mask_cirrus = GeoArray(self.paths.mask_cirrus)[:]
 
+    def transform_swir_to_vnir_raster(self,
+                                      array_swirsensorgeo: np.ndarray,
+                                      swir_lons: np.ndarray,
+                                      swir_lats: np.ndarray,
+                                      swir_epsg: int,
+                                      resamp_alg: str = 'nearest',
+                                      respect_keystone: bool = False
+                                      ) -> np.ndarray:
+        """Transform the given SWIR sensor-geometry raster array into VNIR sensor geometry.
+
+        :param array_swirsensorgeo: source array in SWIR sensor geometry to be transformed
+        :param swir_lons:           longitude geolayer array of the SWIR
+        :param swir_lats:           latitude geolayer array of the SWIR
+        :param swir_epsg:           EPSG code of the SWIR when transformed to map geometry
+        :param resamp_alg:          resampling algorith ('nearest', 'bilinear', 'gauss', 'custom')
+        :param respect_keystone:    whether to use the full geoarray (all bands) in case a 3D array
+                                    in the dimension of the SWIR detector is passed (default: False)
+        """
+        return self._transform_raster_geometry_from_other_detector(
+            array_swirsensorgeo, swir_lons, swir_lats, swir_epsg, resamp_alg, respect_keystone)
+
 
 class EnMAP_SWIR_SensorGeo(EnMAP_Detector_SensorGeo):
     def __init__(self, root_dir: str, config: EnPTConfig, logger=None, meta=None) -> None:
@@ -405,32 +497,22 @@ class EnMAP_SWIR_SensorGeo(EnMAP_Detector_SensorGeo):
                                       array_vnirsensorgeo: np.ndarray,
                                       vnir_lons: np.ndarray,
                                       vnir_lats: np.ndarray,
-                                      vnir_epsg: int) -> np.ndarray:
-        """Set a SWIR raster attribute with a VNIR attribute transformed to SWIR sensor geometry.
+                                      vnir_epsg: int,
+                                      resamp_alg: str = 'nearest',
+                                      respect_keystone: bool = False
+                                      ) -> np.ndarray:
+        """Transform the given VNIR sensor-geometry raster array into SWIR sensor geometry.
 
-        :param array_vnirsensorgeo: source array in VNIR sensor geometry wo be transformed
+        :param array_vnirsensorgeo: source array in VNIR sensor geometry to be transformed
         :param vnir_lons:           longitude geolayer array of the VNIR
         :param vnir_lats:           latitude geolayer array of the VNIR
         :param vnir_epsg:           EPSG code of the VNIR when transformed to map geometry
+        :param resamp_alg:          resampling algorith ('nearest', 'bilinear', 'gauss', 'custom')
+        :param respect_keystone:    whether to use the full geoarray (all bands) in case a 3D array
+                                    in the dimension of the VNIR detector is passed (default: False)
         """
-        if self.detector_meta.lons is None or self.detector_meta.lats is None:
-            raise RuntimeError('The SWIR geolayer must be computed first '
-                               'to transform arrays from VNIR to SWIR sensor geometry.')
-
-        from ...processors.spatial_transform import VNIR_SWIR_SensorGeometryTransformer
-        VS_SGT = VNIR_SWIR_SensorGeometryTransformer(lons_vnir=vnir_lons[:, :, 0],
-                                                     lats_vnir=vnir_lats[:, :, 0],
-                                                     lons_swir=self.detector_meta.lons[:, :, 0],
-                                                     lats_swir=self.detector_meta.lats[:, :, 0],
-                                                     prj_vnir=vnir_epsg,
-                                                     prj_swir=self.detector_meta.epsg_ortho,
-                                                     res_vnir=(30, 30),
-                                                     res_swir=(30, 30),
-                                                     resamp_alg='nearest',
-                                                     # radius_of_influence=45,
-                                                     nprocs=self.cfg.CPUs
-                                                     )
-        return VS_SGT.transform_sensorgeo_VNIR_to_SWIR(array_vnirsensorgeo)
+        return self._transform_raster_geometry_from_other_detector(
+            array_vnirsensorgeo, vnir_lons, vnir_lats, vnir_epsg, resamp_alg, respect_keystone)
 
 
 class EnMAPL1Product_SensorGeo(object):
@@ -699,27 +781,74 @@ class EnMAPL1Product_SensorGeo(object):
         self.vnir.get_preprocessed_dem()
         self.swir.get_preprocessed_dem()
 
-    def transform_vnir_to_swir_raster(self, attrName):
-        """Set the specified SWIR raster attribute with a VNIR attribute transformed to SWIR sensor geometry.
+    def transform_vnir_to_swir_raster(self,
+                                      array_vnirsensorgeo: np.ndarray,
+                                      resamp_alg: str = 'nearest',
+                                      respect_keystone: bool = False
+                                      ) -> np.ndarray:
+        """Transform the given array from VNIR into SWIR sensor geometry.
 
-        :param attrName:    name of the attribute to be set
+        :param array_vnirsensorgeo: raster array in VNIR sensor geometry to be transformed into SWIR sensor geometry
+        :param resamp_alg:          resampling algorithm ('nearest', 'bilinear', 'gauss', 'custom')
+        :param respect_keystone:    whether to use the full geoarray (all bands) in case a 3D array
+                                    in the dimension of the VNIR detector is passed (default: False)
         """
-        self.logger.info("Transforming the '%s' attribute from VNIR to SWIR sensor geometry." % attrName)
-
-        vnir_rasterAttr = getattr(self.vnir, attrName)
-
         if self.meta.vnir.lons is None or self.meta.vnir.lats is None or \
            self.meta.swir.lons is None or self.meta.swir.lats is None:
             raise RuntimeError('The VNIR/SWIR geolayers must be computed first '
                                'to transform arrays from VNIR to SWIR sensor geometry.')
 
+        return self.swir.transform_vnir_to_swir_raster(array_vnirsensorgeo=array_vnirsensorgeo,
+                                                       vnir_lons=self.meta.vnir.lons,
+                                                       vnir_lats=self.meta.vnir.lats,
+                                                       vnir_epsg=self.meta.vnir.epsg_ortho,
+                                                       resamp_alg=resamp_alg,
+                                                       respect_keystone=respect_keystone)
+
+    def transform_swir_to_vnir_raster(self, array_swirsensorgeo: np.ndarray,
+                                      resamp_alg: str = 'nearest',
+                                      respect_keystone: bool = False
+                                      ) -> np.ndarray:
+        """Transform the given array from SWIR into VNIR sensor geometry.
+
+        :param array_swirsensorgeo: raster array in SWIR sensor geometry to be transformed into VNIR sensor geometry
+        :param resamp_alg:          resampling algorithm ('nearest', 'bilinear', 'gauss', 'custom')
+        :param respect_keystone:    whether to use the full geoarray (all bands) in case a 3D array
+                                    in the dimension of the VNIR detector is passed (default: False)
+        """
+        if self.meta.vnir.lons is None or self.meta.vnir.lats is None or \
+           self.meta.swir.lons is None or self.meta.swir.lats is None:
+            raise RuntimeError('The VNIR/SWIR geolayers must be computed first '
+                               'to transform arrays from VNIR to SWIR sensor geometry.')
+
+        return self.vnir.transform_swir_to_vnir_raster(array_swirsensorgeo=array_swirsensorgeo,
+                                                       swir_lons=self.meta.swir.lons,
+                                                       swir_lats=self.meta.swir.lats,
+                                                       swir_epsg=self.meta.swir.epsg_ortho,
+                                                       resamp_alg=resamp_alg,
+                                                       respect_keystone=respect_keystone)
+
+    def set_SWIRattr_with_transformedVNIRattr(self, attrName: str,
+                                              resamp_alg: str = 'nearest',
+                                              respect_keystone: bool = False
+                                              ) -> None:
+        """Set the specified SWIR raster attribute with a VNIR attribute transformed to SWIR sensor geometry.
+
+        :param attrName:            name of the attribute to be set
+        :param resamp_alg:          resampling algorithm ('nearest', 'bilinear', 'gauss', 'custom')
+        :param respect_keystone:    whether to use the full geoarray (all bands) in case the attribute
+                                    to be transformed is 'data' (default: False)
+        """
+        self.logger.info("Transforming the '%s' attribute from VNIR to SWIR sensor geometry." % attrName)
+
+        vnir_rasterAttr = getattr(self.vnir, attrName)
+
         if vnir_rasterAttr is None:
             raise RuntimeError("%s.vnir.%s has not yet been set." % (self.__class__.__name__, attrName))
 
-        attr_transformed = self.swir.transform_vnir_to_swir_raster(array_vnirsensorgeo=np.array(vnir_rasterAttr),
-                                                                   vnir_lons=self.meta.vnir.lons,
-                                                                   vnir_lats=self.meta.vnir.lats,
-                                                                   vnir_epsg=self.meta.vnir.epsg_ortho)
+        attr_transformed = self.transform_vnir_to_swir_raster(array_vnirsensorgeo=np.array(vnir_rasterAttr),
+                                                              resamp_alg=resamp_alg,
+                                                              respect_keystone=respect_keystone)
         setattr(self.swir, attrName, attr_transformed)
 
     def run_AC(self):